+++ js/main.js
@@ -1,101 +1,4 @@
-const maxNumberComments = 100;
-const maxNumberLikes = 200;
-const minNumberLikes = 15;
-const maxNumberAvatar = 6;
+import './utils.js';
+import './data.js';
 
-const DESCRIPTIONS = [
-  'В Милане',
-  'На Кубе',
-  'No war',
-  'Миру - мир',
-  'Не виноватая я, он сам пришел',
-  'По пивку?',
-  'Курю кальян',
-  'Чил',
-  'На работе',
-  'На даче',
-  'Экстемально',
-  'Жара',
-  'Брр, скорее бы лето',
-];
-
-const NAMES = [
-  'Александр',
-  'Павел',
-  'Анастасия',
-  'Яна',
-  'Владимир',
-  'Геннадий',
-  'Юлия',
-  'Вероника',
-  'Алексей',
-  'Андрей',
-  'Полина',
-  'Алина',
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо.',
-  'Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
-  'В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.',
-  'Как можно было поймать такой неудачный момент?!',
-];
-
-const comments = [];
-const posts = [];
-
-function getRandomNumber(min,max) {
-  if (max <= min) {
-    throw('Ошибка');
-  }
-  if (min <= 0 || max <= 0) {
-    throw('Ошибка');
-  }
-  return Math.round(Math.random() * (max - min) + min);
-}
-
-function checkLegth (str, maxLegth) {
-  return str.length > maxLegth;
-}
-
-const getRandomPositiveInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const createComments = () => ({
-  id: getRandomNumber(1, maxNumberComments),
-  avatar: `photos/${getRandomNumber(1, maxNumberAvatar)}.svg.`,
-  message: MESSAGES[getRandomPositiveInteger(0, MESSAGES.length - 1)],
-  name: NAMES[getRandomPositiveInteger(0, NAMES.length - 1)],
-});
-
-const createPosts = (id) => ({
-  id: id,
-  url: `photos/${id}.jpg`,
-  description: DESCRIPTIONS[getRandomPositiveInteger(0, DESCRIPTIONS.length - 1)],
-  likes: getRandomNumber(minNumberLikes, maxNumberLikes),
-  comments: comments,
-});
-
-const postsRandElement = () => {
-  const rand = Math.floor(Math.random() * posts.length);
-  return posts[rand];
-};
-
-const AMTCOMMENTS = 10;
-for (let i = 0; i < AMTCOMMENTS; i++) {
-  comments.push(createComments());
-}
-
-const AMTPOSTS = 25;
-for (let i = 0; i < AMTPOSTS; i++) {
-  posts.push(createPosts(i));
-}
+import {postsRandElement} from './data.js';
